+++ apps/anatolia/babel.config.js
@@ -1,8 +1,9 @@
-/* eslint-disable */
+
+// eslint-disable-next-line no-undef
 module.exports = function async (api) {
   api.cache(true);
   return {
-    presets: ['babel-preset-expo'],
+    presets: ["babel-preset-expo"],
     plugins: ["react-native-reanimated/plugin"],
   };
 };