+++ packages/core/src/components/index.ts
@@ -4,10 +4,10 @@ export { default as Stack } from "./mics/Stack/Stack"
 export { default as Group } from "./mics/Group/Group"
 export { default as Box } from "./mics/Box/Box"
 export { default as Modal } from "./mics/Modal/Modal"
-export { default as BottomSheet } from "./mics/BottomSheet/BottomSheet"
 export { default as View } from "./mics/View/View"
 export { default as FlatList } from "./mics/FlatList/FlatList"
 export { default as ImageBackground } from "./mics/ImageBackground/ImageBackground"
+export { default as Sheet } from "./mics/Sheet/Sheet"
 
 // Typography
 export { default as Text } from "./typography/Text/Text"