+++ packages/core/src/types/SheetRef/SheetRef.ts
@@ -1,0 +1,4 @@
+export type SheetRef = {
+  open: () => void
+  close: () => void
+}